Added throwClassNotFoundOnMissingClassForField (#726)
---------
