+++ src/main/java/net/openhft/chronicle/wire/TextWire.java
@@ -85,7 +85,7 @@ public class TextWire extends AbstractWire implements Wire {
     static final Supplier<StopCharsTester> END_EVENT_NAME_ESCAPING = TextStopCharsTesters.END_EVENT_NAME::escaping;
     static final Bytes<?> META_DATA = Bytes.from("!!meta-data");
     @Deprecated(/* for removal in x.26, make default true in x.25 */)
-    static final boolean IAE_ON_CNF = Jvm.getBoolean("illegal.argument.for.missing.class.alias", false);
+    static final boolean IAE_ON_CNF = Jvm.getBoolean("class.not.found.for.missing.class.alias", false);
 
     static {
         IOTools.unmonitor(TYPE);
@@ -3203,7 +3203,7 @@ public class TextWire extends AbstractWire implements Wire {
                         }
                         String message = "Unable to load " + stringBuilder + ", is a class alias missing.";
                         if (IAE_ON_CNF)
-                            throw new IllegalArgumentException(message);
+                            throw new ClassNotFoundRuntimeException(new ClassNotFoundException(message));
                         Jvm.warn().on(TextWire.class, message);
                         return null;
                     }
@@ -3224,7 +3224,7 @@ public class TextWire extends AbstractWire implements Wire {
                         }
 
                     } else if (tClass.getClassLoader() == null) {
-                        throw new IllegalArgumentException("Unable to find class " + stringBuilder);
+                        throw new ClassNotFoundRuntimeException(new ClassNotFoundException("Unable to find class " + stringBuilder));
                     } else {
                         return Wires.tupleFor(tClass, stringBuilder.toString());
                     }
