+++ system.properties
@@ -29,4 +29,4 @@ wire.generate.v2=false
 chronicle.announcer.disable=true
 dumpCodeToTarget=true
 # to be removed in x.26
-illegal.argument.for.missing.class.alias=true
+class.not.found.for.missing.class.alias=true
